refactor: add key management module and refactor key handling in CLI commands for improved organization
